+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/store/db/DbScanStoreMetrics.scala
@@ -8,7 +8,7 @@ import com.daml.metrics.api.MetricQualification.Latency
 import com.daml.metrics.api.{MetricInfo, MetricName, MetricsContext}
 import com.digitalasset.canton.config.ProcessingTimeout
 import com.digitalasset.canton.lifecycle.{FlagCloseable, LifeCycle, UnlessShutdown}
-import com.digitalasset.canton.logging.TracedLogger
+import com.digitalasset.canton.logging.{NamedLoggerFactory, NamedLogging}
 import com.digitalasset.canton.metrics.CacheMetrics
 import com.digitalasset.canton.tracing.TraceContext
 import org.lfdecentralizedtrust.splice.environment.SpliceMetrics
@@ -16,14 +16,12 @@ import org.lfdecentralizedtrust.splice.store.HistoryMetrics
 
 class DbScanStoreMetrics(
     metricsFactory: LabeledMetricsFactory,
-    val logger: TracedLogger,
+    val loggerFactory: NamedLoggerFactory,
     val timeouts: ProcessingTimeout,
-) extends FlagCloseable {
+) extends FlagCloseable
+    with NamedLogging {
 
-  /** Storing all the created cache metrics to ensure we always have just one instance per name
-    * This is done because creating the same gauge with teh same name is not safe
-    * Eventually we might want to move this all the way into the metrics factory to always return the same gauge
-    */
+  // storing the caches as they have to be closed so that all the created gauges are closed
   private val cacheOfMetrics = scala.collection.concurrent
     .TrieMap[String, CacheMetrics]()
 
@@ -55,8 +53,12 @@ class DbScanStoreMetrics(
       cacheName: String
   )(implicit tc: TraceContext): UnlessShutdown[CacheMetrics] =
     performUnlessClosing(s"register cache $cacheName") {
-      logger.info(s"Registering new cache metrics for ${cacheName}")
-      cacheOfMetrics.getOrElseUpdate(cacheName, new CacheMetrics(cacheName, metricsFactory))
+      cacheOfMetrics.getOrElseUpdate(
+        cacheName, {
+          logger.info(s"Registering new cache metrics for $cacheName")
+          new CacheMetrics(cacheName, metricsFactory)
+        },
+      )
     }
 
   val history = new HistoryMetrics(metricsFactory)(MetricsContext.Empty)