+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/config/SvAppConfig.scala
@@ -205,7 +205,7 @@ final case class SvParticipantClientConfig(
     override val adminApi: FullClientConfig,
     override val ledgerApi: LedgerApiClientConfig,
     sequencerRequestAmplification: SubmissionRequestAmplification =
-      SvAppBackendConfig.DEFAULT_SEQUENCER_REQUEST_AMPLIFICATION,
+      SvAppBackendConfig.DefaultParticipantSequencerRequestAmplification,
 ) extends BaseParticipantClientConfig(adminApi, ledgerApi)
 
 case class SvAppBackendConfig(
@@ -261,6 +261,8 @@ case class SvAppBackendConfig(
       NonNegativeFiniteDuration.ofHours(24),
     // Defaults to 48h as it must be at least 2x preparationTimeRecordtimeTolerance
     mediatorDeduplicationTimeout: NonNegativeFiniteDuration = NonNegativeFiniteDuration.ofHours(48),
+    topologyChangeDelayDuration: NonNegativeFiniteDuration =
+      NonNegativeFiniteDuration.ofMillis(250),
     delegatelessAutomation: Boolean = true,
     expectedTaskDuration: Long = 5000, // milliseconds
     expiredRewardCouponBatchSize: Int = 100,
@@ -284,7 +286,12 @@ case class SvAppBackendConfig(
 }
 
 object SvAppBackendConfig {
-  val DEFAULT_SEQUENCER_REQUEST_AMPLIFICATION = SubmissionRequestAmplification(
+  // This is consistent with what the validator sets for a single sequencer connection
+  val DefaultParticipantSequencerRequestAmplification = SubmissionRequestAmplification(
+    PositiveInt.tryCreate(1),
+    NonNegativeFiniteDuration.ofSeconds(10),
+  )
+  val DefaultMediatorSequencerRequestAmplification = SubmissionRequestAmplification(
     PositiveInt.tryCreate(5),
     NonNegativeFiniteDuration.ofSeconds(10),
   )
@@ -343,7 +350,7 @@ final case class SvSequencerConfig(
 final case class SvMediatorConfig(
     adminApi: FullClientConfig,
     sequencerRequestAmplification: SubmissionRequestAmplification =
-      SvAppBackendConfig.DEFAULT_SEQUENCER_REQUEST_AMPLIFICATION,
+      SvAppBackendConfig.DefaultMediatorSequencerRequestAmplification,
 ) {
 
   def toCantonConfig: RemoteMediatorConfig = RemoteMediatorConfig(